# course-Microservices/catalog-service/openapi.yaml
openapi: 3.0.3
info:
  title: 校园选课系统（单体版）API
  description: |
    校园选课与教学资源管理平台
  version: 1.0.0
  contact:
    name: API Support
    email: support@zjgsu.edu.cn
servers:
  - url: http://localhost:8080
    description: 开发环境
...
paths:
  /api/courses:
    get:
      tags:
        - Course
      summary: 查询所有课程
      operationId: getAllCourses
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseListResp'

    post:
      tags:
        - Course
      summary: 创建课程
      operationId: createCourse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseInput'
            example:
              code: "CS202"
              title: "数据结构"
              instructor:
                id: "T001"
                name: "张教授"
                email: "zhang@zjgsu.edu.cn"
              schedule:
                dayOfWeek: "MONDAY"
                startTime: "08:00"
                endTime: "10:00"
                expectedAttendance: 50
              capacity: 60
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseResp'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/courses/code/{code}:
    get:
      tags:
        - Course
      summary: 根据Code查询课程
      operationId: getCourseByCode
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
            example: "CS202"
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseResp'
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'

    put:
      tags:
        - Course
      summary: 根据Code更新课程
      operationId: updateCourseByCode
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseInput'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseResp'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Course
      summary: 根据Code删除课程
      operationId: deleteCourseByCode
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          $ref: '#/components/responses/DeleteSuccess'
        '404':
          $ref: '#/components/responses/NotFound'

  # ======== Student ========
...
components:
...
  responses:
    BadRequest:
      description: 参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResp'
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResp'
    DeleteSuccess:
      description: 删除成功
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SuccessResp'
  schemas:
    BaseResp:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          nullable: true
          type: object
    ErrorResp:
      allOf:
        - $ref: '#/components/schemas/BaseResp'
      example:
        code: 400
        message: "课程不存在"
        data: null
    SuccessResp:
      allOf:
        - $ref: '#/components/schemas/BaseResp'
      example:
        code: 200
        message: "Deleted"
        data: null

    # ---------- Course ----------
    CourseInput:
      type: object
      required:
        - code
        - title
      properties:
        code:
          type: string
          description: 课程编号
        title:
          type: string
          description: 课程名称
    Course:
      type: object
      properties:
        id:
          type: string
          description: 系统生成的课程主键
        code:
          type: string
        title:
          type: string
    CourseResp:
      allOf:
        - $ref: '#/components/schemas/BaseResp'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Course'
    CourseListResp:
      allOf:
        - $ref: '#/components/schemas/BaseResp'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Course'

    # ---------- Student ----------
...